Change ongoing from javax servlets to WS RS Api

+++ src/main/java/com/iambadatplaying/lcuHandler/SocketClient.java
@@ -52,7 +52,7 @@ public class SocketClient {
 
     public void shutdown() {
         try {
-            client.destroy();
+            if (client != null) client.destroy();
         } catch (Exception e) {
             e.printStackTrace();
         }
